Refactor: ChattingList Folder Location
