+++ app/src/main/java/com/example/geeksasaeng/Chatting/ChattingList/ChatDataView.kt
@@ -1,4 +1,4 @@
-package com.example.geeksasaeng.Chatting.ChattingRoom.ChattingList
+package com.example.geeksasaeng.Chatting.ChattingList
 
 interface ChatDataView {
     fun onSuccessGetChatData(position: Int, chattingData: ChattingData)
