+++ app/src/main/java/com/example/geeksasaeng/Chatting/ChattingList/Retrofit/ChattingListRetrofitInterfaces.kt
@@ -1,4 +1,4 @@
-package com.example.geeksasaeng.Chatting.ChattingRoom.ChattingList.Retrofit
+package com.example.geeksasaeng.Chatting.ChattingList.Retrofit
 
 import retrofit2.Call
 import retrofit2.http.GET
