+++ app/src/main/java/com/example/geeksasaeng/Chatting/ChattingList/Retrofit/ChattingListView.kt
@@ -1,4 +1,4 @@
-package com.example.geeksasaeng.Chatting.ChattingRoom.ChattingList.Retrofit
+package com.example.geeksasaeng.Chatting.ChattingList.Retrofit
 
 interface ChattingListView {
     fun getChattingListSuccess(result: ChattingListResult)
