+++ app/src/main/java/com/example/geeksasaeng/Chatting/ChattingRoom/DialogMatchingEnd.kt
@@ -9,7 +9,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.DialogFragment
-import com.example.geeksasaeng.Chatting.ChattingRoom.ChattingList.ParticipantsInfo
+import com.example.geeksasaeng.Chatting.ChattingList.ParticipantsInfo
 import com.example.geeksasaeng.Chatting.Retrofit.ChattingDataService
 import com.example.geeksasaeng.Chatting.Retrofit.MatchingEndView
 import com.example.geeksasaeng.R
