+++ app/src/main/java/com/example/geeksasaeng/Home/Party/LookParty/DialogPartyRequestComplete.kt
@@ -5,7 +5,7 @@ import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
 import android.view.*
 import androidx.fragment.app.DialogFragment
-import com.example.geeksasaeng.Chatting.ChattingRoom.ChattingList.ChattingListFragment
+import com.example.geeksasaeng.Chatting.ChattingList.ChattingListFragment
 import com.example.geeksasaeng.MainActivity
 import com.example.geeksasaeng.R
 import com.example.geeksasaeng.databinding.DialogPartyRequestCompleteBinding
